+++ delivery/src/main/java/com/nowayback/delivery/infrastructure/deliveryroute/DeliveryRouteJpaRepository.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.infrastructure.repository;
+package com.nowayback.delivery.infrastructure.deliveryroute;
 
 import com.nowayback.delivery.domain.deliveryroute.entity.DeliveryRoute;
 import com.nowayback.delivery.domain.deliveryroute.vo.DeliveryId;
